requirements.txt
gym
numpy==1.16.2
gym-retro
opencv-python
sk-video
visdom
webcolors
tensorflow==1.13.1
